Basic function call frames implemented.
Most of the types I defined the other day are basicall useless for
actual execution, since we have to shuffle them alllll into the actual
store, which breaks every index reference and such, and stuff like storage
for locals goes in different places, and so on.

Oh well, so it goes.