Implemented Select
That was easy; turns out the hard one I was thinking of is BrTable
which I've already implemented (maybe even right).

All opcodes now have an implementation, even if it's an incorrect one.
I need to go through and do a big pile of cleanup, and then I think the
next step will be to integrate the official wasm unit tests so I can
actually find the instructions that are wrong.