Fixed ONE bug with start functions in test code
The created test function was marked as the start function
(which is called `main` in parity-wasm, not start, conveniently),
which is not valid 'cause the start function can't take parameters.

Now there's still a bug when, I think, a start function returns it
tries to pop its stack frame and there's no stack frame beneath it
or such.  Something like that at least.